requirements.txt
amqp==2.5.1
Babel==2.7.0
billiard==3.6.1.0
celery==4.3.0
Django==2.2.4
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-extensions==2.2.1
django-phone-field==1.8.0
django-phonenumber-field==3.0.1
django-timezone-field==3.0
importlib-metadata==0.19
kombu==4.6.4
more-itertools==7.2.0
phonenumbers==8.10.17
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.2
red==0.2.2
redis==3.3.8
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
amqp==2.5.1
Babel==2.7.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-extensions==2.2.1
django-phone-field==1.8.0
django-phonenumber-field==3.0.1
django-timezone-field==3.0
idna==2.8
importlib-metadata==0.19
kombu==4.6.4
more-itertools==7.2.0
phonenumbers==8.10.17
PyJWT==1.7.1
PySocks==1.7.0
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.2
red==0.2.2
redis==3.3.8
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
twilio==6.29.4
urllib3==1.25.3
vine==1.3.0
zipp==0.6.0
amqp==2.5.1
Babel==2.7.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
darkskylib==0.3.91
Django==2.2.4
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-extensions==2.2.1
django-phone-field==1.8.0
django-phonenumber-field==3.0.1
django-timezone-field==3.0
future==0.17.1
idna==2.8
importlib-metadata==0.19
kombu==4.6.4
more-itertools==7.2.0
phonenumbers==8.10.17
PyJWT==1.7.1
PySocks==1.7.0
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.2
red==0.2.2
redis==3.3.8
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
twilio==6.29.4
urllib3==1.25.3
vine==1.3.0
zipp==0.6.0
